InstitutionDicom moved to commons + first draft of findCenterOrCreateByInstitutionDicom in ms studies
