+++ shanoir-ng-ms-common/src/main/java/org/shanoir/ng/shared/dicom/InstitutionDicom.java
@@ -12,10 +12,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
  */
 
-/**
- * 
- */
-package org.shanoir.ng.importer.model;
+package org.shanoir.ng.shared.dicom;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
